FEAT: Added two functions in src/ib.rs
I was looking or a quick and simple way to get Configured IP addresses
to be able to make a SocketAddr to bind a TcpListener and found this
crate was useful but I was missing a way to specify the network
interface and a way to check user provided interface names against the
list of all network interfaces on the system to be able to perform some
error handling if the user misspell the interface name or give a wrong
one.

* First function : pub fn get_ifaces() -> Vec<String>

This function takes no arguments and returns a Vec of all networks
interfaces using `ls /sys/class/net` to stay in the philosophy of the
crate.

* Second function pub fn get_iface_addr(iface: &str) -> Option<IpAddr>

This function takes a &str of the name of a network interface name  and
returns the IpAddr of the IP address of that interface.